+++ access_roles/__init__.py
@@ -20,3 +20,4 @@
 #
 #############################################################################
 from . import models
+from .hooks import post_init_hook