+++ access_roles/models/res_groups.py
@@ -54,12 +54,6 @@ class ResGroups(models.Model):
         """Extract group IDs from a selection group field name."""
         return [int(v) for v in name[11:].split('_')]
 
-    @api.model
-    def _register_hook(self):
-        """Hook to update role-based group view after module installation."""
-        super()._register_hook()
-        self._update_role_groups_view()
-        return True
 
     @api.model
     def get_groups_by_application(self):